notify/pyproject.toml
[tool.poetry]
name = "notifyms"
version = "0.1.0"
description = ""
authors = ["ZheTao <zacharytao99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pika = "^1.3.1"
responses = "^0.23.1"
sendgrid = "^6.10.0"
python-dotenv = "^1.0.0"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
